Refactor visualization functions and enhance layer dimension retrieval in utility modules
